// src/utils/avatarUrl.d.ts
/**
 * Utility function to convert relative avatar URL to absolute URL
 * @param {Object} req - Express request object
 * @param {string} avatarUrl - Avatar URL (could be relative or absolute)
 * @returns {string} Absolute avatar URL
 */
declare const getAbsoluteAvatarUrl: (req: any, avatarUrl: any) => any;